requirements.txt
argh==0.26.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.5.14
cssmin==0.2.0
cssprefixer==1.4.1
cssutils==1.0
Flask==0.10.1
Flask-Assets==0.10
Flask-Bootstrap==3.3.0.1
Flask-FlatPages==0.6
Flask-Moment==0.4.0
Flask-Script==2.0.5
Frozen-Flask==0.11
itsdangerous==0.24
Jinja2==2.7.3
jsmin==2.1.1
livereload==2.3.2
Markdown>2.6
MarkupSafe==0.23
pathtools==0.1.2
PyYAML==3.11
sass==2.3
tornado==4.1
watchdog==0.8.3
webassets==0.10.1
Werkzeug==0.10.1
